Added jwt verification and cleaned up res/req dtos

+++ src/Api/Domain/Users/Controllers/UserController.cs
@@ -2,14 +2,16 @@ using Api.Domain.Users.DTOs;
 using Api.Domain.Users.Models;
 using Api.Domain.Users.Services;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Domain.Users.Controllers;
 
 [ApiController]
-[Route("api/[controller]")]
-public class UsersController(IUserService service, IMapper mapper) : ControllerBase
+[Route("api/users")]
+public class UserController(IUserService service, ICurrentUserService currentUserService, IMapper mapper) : ControllerBase
 {
+	[Authorize]
     [HttpGet]
     public async Task<ActionResult<ICollection<GetUserResponseDto>>> GetAllUsers()
     {
@@ -17,6 +19,7 @@ public class UsersController(IUserService service, IMapper mapper) : ControllerB
         return Ok(mapper.Map<ICollection<GetUserResponseDto>>(users));
     }
 
+	[Authorize]
     [HttpGet("{userId}")]
     public async Task<ActionResult<GetUserResponseDto>> GetUser(Guid userId)
     {
@@ -24,33 +27,40 @@ public class UsersController(IUserService service, IMapper mapper) : ControllerB
         return Ok(mapper.Map<GetUserResponseDto>(user));
     }
 
+	[Authorize]
     [HttpGet("search")]
     public async Task<ActionResult<GetUserResponseDto>> GetUser([FromQuery] string email, [FromQuery] string username)
     {
         if(string.IsNullOrEmpty(email) && string.IsNullOrEmpty(username))
             return BadRequest("Either email or username must be provided.");
         
-        User user = string.IsNullOrEmpty(email) 
+        var user = string.IsNullOrEmpty(email) 
             ? await service.GetByUsernameAsync(username)
             : await service.GetByEmailAsync(email);
-
+		
         return Ok(mapper.Map<GetUserResponseDto>(user));
     }
 
-    //[Authorize]
+	[Authorize]
     [HttpPatch("{userId}")]
     public async Task<ActionResult<UpdateUserResponseDto>> UpdateUser(Guid userId, [FromBody] UpdateUserRequestDto dto)
     {
+		if(currentUserService.GetCurrentUserId() != userId)
+            return Unauthorized("You can only update your own account.");
+
         await service.CheckPasswordAsync(userId, dto.Password.Trim());
         var updatedUser = await service.UpdateAsync(userId, dto);
         return Ok(mapper.Map<UpdateUserResponseDto>(updatedUser));
     }
 
-    //[Authorize]
+	[Authorize]
     [HttpDelete("{userId}")]
     public async Task<ActionResult> DeleteUser(Guid userId, [FromBody] DeleteUserRequestDto dto)
     {
-        //await service.CheckPasswordAsync(userId, dto.Password.Trim());
+		if(currentUserService.GetCurrentUserId() != userId)
+            return Unauthorized("You can only delete your own account.");
+
+        await service.CheckPasswordAsync(userId, dto.Password.Trim());
         await service.DeleteAsync(userId);
         return Ok();
     }
